Refactor Content Helper Error Message code
