+++ src/blocks/content-helper/content-helper.tsx
@@ -12,7 +12,7 @@ import { registerPlugin } from '@wordpress/plugins';
 import PerformanceDetails from './performance-details/component';
 import RelatedTopPostList from './related-top-posts/component-list';
 import LeafIcon from '../shared/components/leaf-icon';
-import { ElementOrEmptyCredentialsMessage } from './content-helper-error';
+import VerifyCredentials from '../../content-helper/verify-credentials';
 
 const BLOCK_PLUGIN_ID = 'wp-parsely-block-editor-sidebar';
 
@@ -20,12 +20,20 @@ const renderSidebar = () => (
 	<PluginSidebar icon={ <LeafIcon /> } name="wp-parsely-content-helper" className="wp-parsely-content-helper" title={ __( 'Parse.ly Editor Sidebar', 'wp-parsely' ) }>
 		<Panel>
 			<PanelBody title={ __( 'Performance Details', 'wp-parsely' ) } initialOpen={ true }>
-				{ ElementOrEmptyCredentialsMessage( <PerformanceDetails /> ) }
+				{
+					<VerifyCredentials>
+						<PerformanceDetails />
+					</VerifyCredentials>
+				}
 			</PanelBody>
 		</Panel>
 		<Panel>
 			<PanelBody title={ __( 'Related Top Posts', 'wp-parsely' ) } initialOpen={ false }>
-				{ ElementOrEmptyCredentialsMessage( <RelatedTopPostList /> ) }
+				{
+					<VerifyCredentials>
+						<RelatedTopPostList />
+					</VerifyCredentials>
+				}
 			</PanelBody>
 		</Panel>
 	</PluginSidebar>
